# docs/openapi.yaml
openapi: 3.0.3
info:
  title: Linkedin-Hyper API
  version: 1.0.0
  description: >-
    Browser-facing Next.js BFF API for Linkedin-Hyper. This spec documents the
    operator dashboard routes, account/session flows, inbox APIs, send/sync APIs,
    health endpoints, export endpoints, and supporting account/stats endpoints.
servers:
  - url: /
    description: Same-origin frontend deployment
  - url: http://127.0.0.1:3002
    description: Local frontend container or dev server
  - url: http://localhost:3000
    description: Local Next.js dev server
...
paths:
...
  /api/inbox/unified:
    get:
      tags: [Inbox]
      summary: Get the unified inbox
      security:
        - cookieAuth: []
        - serviceToken: []
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 200
        - name: offset
          in: query
          schema:
            type: integer
            minimum: 0
      responses:
        '200':
          description: Unified inbox payload
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/InboxUnifiedResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '502':
          description: Backend unreachable
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
  /api/messages/send:
    post:
      tags: [Messages]
      summary: Deprecated send route
      deprecated: true
      security:
        - cookieAuth: []
        - serviceToken: []
      responses:
        '410':
          description: Deprecated route
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DeprecatedRouteResponse'
  /api/messages/send-new:
    post:
      tags: [Messages]
      summary: Send a LinkedIn message
      description: >-
        Sends a controlled message through the worker. Requires account access,
        message text, and either profileUrl or chatId.
      security:
        - cookieAuth: []
        - serviceToken: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SendNewRequest'
      responses:
        '200':
          description: Message send accepted/completed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SendNewResponse'
        '400':
          description: Validation failed
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '429':
          description: Cooldown or rate limit applied
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/OperationErrorResponse'
...
  /api/people/search:
    get:
      tags: [Connections]
      summary: Search people through the worker
      security:
        - cookieAuth: []
        - serviceToken: []
      parameters:
        - name: q
          in: query
          required: false
          schema:
            type: string
      responses:
        '200':
          description: Search results returned
          content:
            application/json:
              schema:
                type: object
                additionalProperties: true
        '401':
          $ref: '#/components/responses/Unauthorized'
...
components:
  securitySchemes:
    cookieAuth:
      type: apiKey
      in: cookie
      name: app_session
      description: Dashboard session cookie.
    apiKeyAuth:
      type: apiKey
      in: header
      name: x-api-key
      description: Controlled operator path for specific worker/admin recovery routes only.
    serviceToken:
      type: http
      scheme: bearer
      bearerFormat: opaque
      description: Compatibility-only service bearer token path for operator automation.
...
  responses:
    Unauthorized:
      description: Request is not authenticated
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
    Forbidden:
      description: Request is authenticated but not authorized or same-origin checks failed
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
  schemas:
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
        code:
          type: string
          nullable: true
        retryAfterSec:
          type: integer
          nullable: true
...
    Message:
      type: object
      required: [id, text, sentAt, sentByMe, senderName]
      properties:
        id:
          type: string
        text:
          type: string
        sentAt:
          type: integer
          format: int64
        sentByMe:
          type: boolean
        senderName:
          type: string
        status:
          type: string
          enum: [sending, sent, failed]
          nullable: true
        error:
          type: string
          nullable: true
    ConversationParticipant:
      type: object
      required: [name, profileUrl]
      properties:
        name:
          type: string
        profileUrl:
          type: string
        avatarUrl:
          type: string
          nullable: true
    ConversationLastMessage:
      type: object
      required: [text, sentAt, sentByMe]
      properties:
        text:
          type: string
        sentAt:
          type: integer
          format: int64
        sentByMe:
          type: boolean
        status:
          type: string
          enum: [sending, sent, failed]
          nullable: true
    Conversation:
      type: object
      required: [conversationId, accountId, participant, lastMessage, unreadCount, messages]
      properties:
        conversationId:
          type: string
        accountId:
          type: string
        participant:
          $ref: '#/components/schemas/ConversationParticipant'
        lastMessage:
          $ref: '#/components/schemas/ConversationLastMessage'
        unreadCount:
          type: integer
        messages:
          type: array
          items:
            $ref: '#/components/schemas/Message'
    InboxUnifiedResponse:
      type: object
      required: [conversations]
      properties:
        conversations:
          type: array
          items:
            $ref: '#/components/schemas/Conversation'
        stale:
          type: boolean
          nullable: true
        degraded:
          type: boolean
          nullable: true
    DeprecatedRouteResponse:
      type: object
      required: [error, code]
      properties:
        error:
          type: string
        code:
          type: string
          enum: [SEND_ROUTE_DEPRECATED]
    SendNewRequest:
      type: object
      required: [accountId, text]
      properties:
        accountId:
          type: string
        profileUrl:
          type: string
          description: LinkedIn profile URL. Required if chatId is not provided.
        chatId:
          type: string
          description: Existing conversation/thread identifier. Required if profileUrl is not provided.
        text:
          type: string
          minLength: 1
          maxLength: 3000
    SendNewResponse:
      type: object
      properties:
        id:
          type: string
          nullable: true
        chatId:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
      additionalProperties: true
...
    OperationErrorResponse:
      allOf:
        - $ref: '#/components/schemas/ErrorResponse'
        - type: object
          properties:
            retryAfterSec:
              type: integer
              nullable: true
...